requirements.txt
Django>=4.2.16,<5.0
djangorestframework>=3.14.0,<4.0
python-dotenv>=1.0.0,<2.0
whitenoise>=6.6.0,<7.0
gunicorn>=21.2.0,<22.0
django-filter>=23.5,<24.0
Pillow>=10.0.0,<11.0
pandas>=2.1.0,<3.0
uvicorn>=0.24.0,<1.0
django-debug-toolbar>=4.2.0,<5.0
sentry-sdk>=1.35.0,<2.0
django-cors-headers>=4.3.0,<5.0
django-cacheops>=7.0.1,<8.0
django-redis>=5.4.0,<6.0
orjson>=3.9.10,<4.0
aiosqlite>=0.19.0,<1.0
watchfiles>=0.21.0,<1.0